sanjeevani-2.0/sanjeevani-backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
motor==3.1.2
pydantic==1.10.8
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0.post2
python-dotenv==1.0.0
firebase-admin==6.1.0
twilio==8.2.2
requests==2.31.0
aiohttp==3.8.4
python-socketio==5.8.0
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.1
sentry-sdk==1.28.1
